Move visualization into detail
