+++ context/app/static/js/components/Detail/visualization/Visualization/style.js
