+++ context/app/static/js/components/Detail/visualization/VisualizationThemeSwitch/VisualizationThemeSwitch.jsx
