+++ context/app/static/js/components/Detail/visualization/VisualizationThemeSwitch/index.js
